Simplified EL support in LoginToContinue
