+++ impl/src/main/java/org/glassfish/soteria/cdi/LoginToContinueAnnotationLiteral.java
@@ -39,6 +39,9 @@
  */
 package org.glassfish.soteria.cdi;
 
+import static org.glassfish.soteria.cdi.AnnotationELPProcessor.evalELExpression;
+import static org.glassfish.soteria.cdi.AnnotationELPProcessor.evalImmediate;
+
 import javax.enterprise.util.AnnotationLiteral;
 import javax.security.enterprise.authentication.mechanism.http.LoginToContinue;
 
@@ -53,16 +56,40 @@ public class LoginToContinueAnnotationLiteral extends AnnotationLiteral<LoginToC
     private final String loginPage;
     private final boolean useForwardToLogin;
     private final String errorPage;
+    
+    private boolean hasDeferredExpressions;
 
     public LoginToContinueAnnotationLiteral(String loginPage, boolean useForwardToLogin, String errorPage) {
         this.loginPage = loginPage;
         this.useForwardToLogin = useForwardToLogin;
         this.errorPage = errorPage;
     }
+    
+    public static LoginToContinue eval(LoginToContinue in) {
+        if (!hasAnyELExpression(in)) {
+            return in;
+        }
+        
+        LoginToContinueAnnotationLiteral out =
+            new LoginToContinueAnnotationLiteral(
+                    evalImmediate(in.loginPage()), 
+                    in.useForwardToLogin(), 
+                    evalImmediate(in.errorPage()));
+        
+        out.setHasDeferredExpressions(hasAnyELExpression(out));
+        
+        return out;
+    }
+    
+    public static boolean hasAnyELExpression(LoginToContinue in) {
+        return AnnotationELPProcessor.hasAnyELExpression(
+            in.loginPage(), 
+            in.errorPage());
+    }
 
     @Override
     public String loginPage() {
-        return loginPage;
+        return hasDeferredExpressions? evalELExpression(loginPage) : loginPage;
     }
 
     @Override
@@ -72,6 +99,14 @@ public class LoginToContinueAnnotationLiteral extends AnnotationLiteral<LoginToC
 
     @Override
     public String errorPage() {
-        return errorPage;
+        return hasDeferredExpressions? evalELExpression(errorPage) : errorPage;
+    }
+    
+    public boolean isHasDeferredExpressions() {
+        return hasDeferredExpressions;
+    }
+
+    public void setHasDeferredExpressions(boolean hasDeferredExpressions) {
+        this.hasDeferredExpressions = hasDeferredExpressions;
     }
 }
